Move logic to lib
